HTRACE-180 Move the GUI to a top-level subproject. (stack via cmccabe)
Create new module named htrace-webapp to house webapp. Add a pom to it
that generates a WAR file. Move the web dir from htrace-htraced to this
new module at src/main/web Added a web.xml so WAR builds. Changed the
symlink over in htrace-htraced/go/gobuild.sh to point at the new
location.  Fixed a bug where the java REST client was looking for
htraced in the wrong location.